+++ CHANGELOG.md
@@ -7,11 +7,16 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+### Added
+
+- Add more tests for `deserialize_message` function ([#61])
+
 ### Changed
 
 - Fix handling parameters whose names are reserved by Python ([#56])
 
 [#56]: https://github.com/openlawlibrary/pygls/pull/56
+[#61]: https://github.com/openlawlibrary/pygls/pull/61
 
 ## [0.7.4] - 03/21/2019