+++ CHANGELOG.md
@@ -13,10 +13,12 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Changed
 
+- Response object should contain result OR error field ([#64])
 - Fix handling parameters whose names are reserved by Python ([#56])
 
-[#56]: https://github.com/openlawlibrary/pygls/pull/56
+[#64]: https://github.com/openlawlibrary/pygls/pull/64
 [#61]: https://github.com/openlawlibrary/pygls/pull/61
+[#56]: https://github.com/openlawlibrary/pygls/pull/56
 
 ## [0.7.4] - 03/21/2019